# backend/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Quotes-Parts - Authentification",
        "description": "Documentation des routes d’authentification (login, logout, me).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/regies": {
            "get": {
                "tags": [
                    "Regies"
                ],
                "summary": "Lister toutes les régies",
                "description": "Retourne la liste complète des régies.",
                "operationId": "1a84b8b3ff96d3921737fc6004eb5909",
                "responses": {
                    "200": {
                        "description": "Liste des régies récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Regie"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Regies"
                ],
                "summary": "Créer une nouvelle régie",
                "description": "Ajoute une nouvelle régie.",
                "operationId": "62115ef656fe875e525a48fb94656aca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "REG_LIBELLE",
                                    "REG_SIGLE",
                                    "REG_SIGLE_CODE"
                                ],
                                "properties": {
                                    "REG_LIBELLE": {
                                        "type": "string"
                                    },
                                    "REG_SIGLE": {
                                        "type": "string"
                                    },
                                    "REG_SIGLE_CODE": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Régie créée avec succès"
                    },
                    "409": {
                        "description": "Une régie avec ce nom ou sigle existe déjà"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/regies/{code}": {
            "get": {
                "tags": [
                    "Regies"
                ],
                "summary": "Afficher une régie",
                "description": "Retourne les détails d’une régie à partir de son code.",
                "operationId": "1e6e5be9030dd540cdb316f1be7c4c75",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code de la régie",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Régie trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Regie"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Régie non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Regies"
                ],
                "summary": "Mettre à jour une régie",
                "description": "Modifie les informations d’une régie existante.",
                "operationId": "0b3a5c8ec424f2d0ddecd86665659eb7",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code de la régie à modifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "REG_LIBELLE": {
                                        "type": "string"
                                    },
                                    "REG_SIGLE": {
                                        "type": "string"
                                    },
                                    "REG_SIGLE_CODE": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Régie mise à jour avec succès"
                    },
                    "404": {
                        "description": "Régie non trouvée"
                    },
                    "409": {
                        "description": "Une régie avec ce nom ou sigle existe déjà"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Regies"
                ],
                "summary": "Supprimer une régie",
                "description": "Supprime une régie existante par son code.",
                "operationId": "a62a2364487f7078e0fa897b3945a2c9",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code de la régie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Régie supprimée avec succès"
                    },
                    "404": {
                        "description": "Régie non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Regie": {
                "title": "Regie",
                "description": "Représentation d’une régie",
                "properties": {
                    "REG_CODE": {
                        "type": "string"
                    },
                    "REG_LIBELLE": {
                        "type": "string"
                    },
                    "REG_SIGLE": {
                        "type": "string"
                    },
                    "REG_SIGLE_CODE": {
                        "type": "string"
                    },
                    "REG_STATUT": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "REG_DATE_CREER": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "REG_DATE_MODIFIER": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "REG_VERSION": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}